Use range for all aws-cdk deps
This is needed so that we resolve the same version of CDK in
both dev deps and normal deps.

It might be that we want to change this to freeze it by an exact
version instead of opening up like this. It depends a bit how
the effect actually turns out to be, and how it will work together
with consumers of this package.

Not sure if this Renovate config is correct, but we'll find out.