fix: move cdk packages to {peer,dev} dependencies
This makes the host application/consumers responsible for controlling the
version of CDK used, so that we avoid multiple versions in the dependency
tree.

This might cause issues when we take use of new CDK packages, as the consumer
applications will also need to add these dependencies explicitly. So
strictly that will be a breaking change. Eventually this will hopefully
be solved if/when CDK moves to monolithic packaging.

We need to keep it in dev dependencies so that we can build and test the code.