chore: fix issues with eslint when having dot in dirname
For a so far unknown reason, eslint fails when we build a branch
that contains a dot, which happens e.g. due to Renovate branches.
It seems that the problem might be related to the include statement
in tsconfig, and that the override of this through tsconfig.eslint.json
doesn't work as expected.

Removing the override and instead having a common base that does not
specify the include seem to solve this.

I have been unable to reproduce this locally, it only happens
in Jenkins.