feat(alarms): add option to disable OK action
It doesn't mean much with an OK action on an error log with evaluation period 1.
You only double the amount of messages.
The fact that the system stopped logging is evident by the lack of more alarm messages.
Otherwise, your system logs alarms constantly every minute and this should be evident in dashboards
or elsewhere.