fix: replace usage of deprecated enum with hardcoded value
In newer versions of CDK there's a new enum ssm.ParameterValueType that
we could use. This would, however, require us to update peer
dependencies to a much stricter version range for CDK. As a
workaround we instead hardcode the necessary value. In the future, as
most consumers have moved on to newer versions of CDK, we'll likely want
to replace the hardcoded value with ssm.ParameterValueType.STRING.