Fix loading of authorizer lambda code in API Gateway construct
When using `NodejsFunction` with `entry`, we point CDK to a file with
our lambda code. This can either be a TypeScript or JavaScript file.
When creating constructs in tests inside this library, the entry will
point to the TypeScript file in the source code. But when a library
consumer uses this, it will instead point to the transpiled JavaScript
file. So to set the correct file extension, we must check if we're being
run in the context of the TypeScript source code (src/api-gateway),
otherwise we're being run by a consumer as transpiled JavaScript.