feat(api-gw): support alternate credential secret format
In the current ApiGateway implementation, basic auth credentials must be
stored on the following format in Secrets Manager in order for the basic
auth authorizer to parse it:
```json
{ "username": "<username>", "password": "<password>" }
```

However, some projects instead store credentials on the following
format:
```json
{ "credentials": "[\"<encoded-credential-1>\",\"<encoded-credential-2>\"]" }
```
...consisting of:
- A single key, `credentials`
- With a _string_ value
- Which is a stringified, escaped JSON array of base64-encoded
  credentials
- In which each element is encoded from `<username>:<password>`

The reason that this second format stores stringified JSON _inside_
JSON, is due to a limitation in Liflig's `load-secrets` library, which
only allows storing strings values.

This commit changes the basic auth Lambda authorizer to allow
credential secrets on this second format. Tests have been expanded to
cover this case for all authorizers.