feat(ecs): add open telemetry sidecar construct (#140)
This will create a sidecar on a FargateService construct,
which collects open telemetry-data from the service's instrumentation
agent.
The data is forwarded to CloudWatch Metrics and XRay.

Renovate is changed to update the docker image used for adot, but also disable automerge for it
so we can manually validate that the collector's config file still works.